+++ banyand/kv/badger.go
@@ -26,9 +26,9 @@ import (
 
 	"github.com/dgraph-io/badger/v3"
 	"github.com/dgraph-io/badger/v3/banyandb"
+	"github.com/dgraph-io/badger/v3/skl"
 	"github.com/dgraph-io/badger/v3/y"
 
-	"github.com/apache/skywalking-banyandb/banyand/observability"
 	"github.com/apache/skywalking-banyandb/pkg/encoding"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
 )
@@ -49,8 +49,8 @@ type badgerTSS struct {
 	dbOpts badger.Options
 }
 
-func (b *badgerTSS) Stats() (s observability.Statistics) {
-	return badgerStats(b.db)
+func (b *badgerTSS) Handover(skl *skl.Skiplist) error {
+	return b.db.HandoverIterator(skl.NewUniIterator(false))
 }
 
 func (b *badgerTSS) Close() error {
@@ -58,14 +58,6 @@ func (b *badgerTSS) Close() error {
 		return b.db.Close()
 	}
 	return nil
-}
-
-func badgerStats(db *badger.DB) (s observability.Statistics) {
-	stat := db.Stats()
-	return observability.Statistics{
-		MemBytes:    stat.MemBytes,
-		MaxMemBytes: db.Opts().MemTableSize,
-	}
 }
 
 type mergedIter struct {
@@ -121,10 +113,6 @@ func (i mergedIter) Value() y.ValueStruct {
 type badgerDB struct {
 	db     *badger.DB
 	dbOpts badger.Options
-}
-
-func (b *badgerDB) Stats() observability.Statistics {
-	return badgerStats(b.db)
 }
 
 func (b *badgerDB) Scan(prefix, seekKey []byte, opt ScanOpts, f ScanFunc) error {