+++ banyand/observability/meter_prom.go
@@ -1,3 +1,6 @@
+//go:build prometheus
+// +build prometheus
+
 // Licensed to Apache Software Foundation (ASF) under one or more contributor
 // license agreements. See the NOTICE file distributed with
 // this work for additional information regarding copyright
@@ -15,9 +18,14 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Package observability provides metrics, profiling, and etc.
 package observability
 
-import "errors"
+import (
+	"github.com/apache/skywalking-banyandb/pkg/meter"
+	"github.com/apache/skywalking-banyandb/pkg/meter/prom"
+)
 
-var errNoAddr = errors.New("no address")
+// NewMeterProvider returns a meter.Provider based on the given scope.
+func NewMeterProvider(scope meter.Scope) meter.Provider {
+	return prom.NewProvider(scope)
+}