+++ banyand/tsdb/buffer.go
@@ -21,16 +21,38 @@ import (
 	"fmt"
 	"sync"
 	"time"
+	"unsafe"
 
 	"github.com/dgraph-io/badger/v3/skl"
 	"github.com/dgraph-io/badger/v3/y"
 
+	"github.com/apache/skywalking-banyandb/api/common"
+	"github.com/apache/skywalking-banyandb/banyand/observability"
 	"github.com/apache/skywalking-banyandb/pkg/convert"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/meter"
 	"github.com/apache/skywalking-banyandb/pkg/run"
 )
 
-const defaultSize = 1 << 20 // 1MB
+const (
+	defaultSize = 1 << 20 // 1MB
+	nodeAlign   = int(unsafe.Sizeof(uint64(0))) - 1
+)
+
+var (
+	bufferMeterProvider meter.Provider
+	maxBytes            meter.Gauge
+	immutableBytes      meter.Gauge
+	mutableBytes        meter.Gauge
+)
+
+func init() {
+	bufferMeterProvider = observability.NewMeterProvider(meterTSDB.SubScope("buffer"))
+	labelNames := append(common.LabelNames(), "bucket")
+	maxBytes = bufferMeterProvider.Gauge("max_bytes", labelNames...)
+	mutableBytes = bufferMeterProvider.Gauge("mutable_bytes", labelNames...)
+	immutableBytes = bufferMeterProvider.Gauge("immutable_bytes", labelNames...)
+}
 
 type operation struct {
 	key   []byte
@@ -52,9 +74,9 @@ type bufferShardBucket struct {
 	flushWaitGroup *sync.WaitGroup
 	log            *logger.Logger
 	immutables     []*skl.Skiplist
+	labelValues    []string
 	index          int
-	flushSize      int
-	size           int
+	capacity       int
 	mutex          sync.RWMutex
 }
 
@@ -71,11 +93,7 @@ type Buffer struct {
 }
 
 // NewBuffer creates a new Buffer instance with the given parameters.
-func NewBuffer(log *logger.Logger, flushSize, writeConcurrency, numShards int, onFlushFn onFlush) (*Buffer, error) {
-	size := flushSize
-	if size < defaultSize {
-		size = defaultSize
-	}
+func NewBuffer(log *logger.Logger, position common.Position, flushSize, writeConcurrency, numShards int, onFlushFn onFlush) (*Buffer, error) {
 	buckets := make([]bufferShardBucket, numShards)
 	buffer := &Buffer{
 		buckets:     buckets,
@@ -89,23 +107,24 @@ func NewBuffer(log *logger.Logger, flushSize, writeConcurrency, numShards int, o
 	for i := 0; i < numShards; i++ {
 		buckets[i] = bufferShardBucket{
 			index:          i,
-			size:           size,
-			mutable:        skl.NewSkiplist(int64(size)),
-			flushSize:      flushSize,
+			capacity:       flushSize,
+			mutable:        skl.NewSkiplist(int64(flushSize)),
 			writeCh:        make(chan operation, writeConcurrency),
 			flushCh:        make(chan flushEvent, 1),
 			writeWaitGroup: &buffer.writeWaitGroup,
 			flushWaitGroup: &buffer.flushWaitGroup,
 			log:            buffer.log.Named(fmt.Sprintf("shard-%d", i)),
+			labelValues:    append(position.LabelValues(), fmt.Sprintf("%d", i)),
 		}
 		buckets[i].start(onFlushFn)
+		maxBytes.Set(float64(flushSize), buckets[i].labelValues...)
 	}
 	return buffer, nil
 }
 
 // Write adds a key-value pair with a timestamp to the appropriate shard bucket in the buffer.
 func (b *Buffer) Write(key, value []byte, timestamp time.Time) {
-	if !b.entryCloser.AddRunning() {
+	if b == nil || !b.entryCloser.AddRunning() {
 		return
 	}
 	defer b.entryCloser.Done()
@@ -119,10 +138,16 @@ func (b *Buffer) Write(key, value []byte, timestamp time.Time) {
 
 // Read retrieves the value associated with the given key and timestamp from the appropriate shard bucket in the buffer.
 func (b *Buffer) Read(key []byte, ts time.Time) ([]byte, bool) {
+	if b == nil || !b.entryCloser.AddRunning() {
+		return nil, false
+	}
+	defer b.entryCloser.Done()
 	keyWithTS := y.KeyWithTs(key, uint64(ts.UnixNano()))
 	index := b.getShardIndex(key)
 	epoch := uint64(ts.UnixNano())
-	for _, bk := range b.buckets[index].getAll() {
+	ll, deferFn := b.buckets[index].getAll()
+	defer deferFn()
+	for _, bk := range ll {
 		value := bk.Get(keyWithTS)
 		if value.Meta == 0 && value.Value == nil {
 			continue
@@ -161,7 +186,7 @@ func (b *Buffer) getShardIndex(key []byte) uint64 {
 	return convert.Hash(key) % uint64(b.numShards)
 }
 
-func (bsb *bufferShardBucket) getAll() []*skl.Skiplist {
+func (bsb *bufferShardBucket) getAll() ([]*skl.Skiplist, func()) {
 	bsb.mutex.RLock()
 	defer bsb.mutex.RUnlock()
 	allList := make([]*skl.Skiplist, len(bsb.immutables)+1)
@@ -172,7 +197,11 @@ func (bsb *bufferShardBucket) getAll() []*skl.Skiplist {
 		allList[i+1] = bsb.immutables[last-i]
 		bsb.immutables[last-i].IncrRef()
 	}
-	return allList
+	return allList, func() {
+		for _, l := range allList {
+			l.DecrRef()
+		}
+	}
 }
 
 func (bsb *bufferShardBucket) start(onFlushFn onFlush) {
@@ -180,34 +209,56 @@ func (bsb *bufferShardBucket) start(onFlushFn onFlush) {
 		defer bsb.flushWaitGroup.Done()
 		for event := range bsb.flushCh {
 			oldSkipList := event.data
-			if err := onFlushFn(bsb.index, oldSkipList); err != nil {
-				bsb.log.Err(err).Msg("flushing immutable buffer failed")
-				continue
+			memSize := oldSkipList.MemSize()
+			for {
+				if err := onFlushFn(bsb.index, oldSkipList); err != nil {
+					bsb.log.Err(err).Msg("flushing immutable buffer failed. Retrying...")
+					time.Sleep(time.Second)
+					continue
+				}
+				break
 			}
 			bsb.mutex.Lock()
 			bsb.immutables = bsb.immutables[1:]
-			oldSkipList.DecrRef()
 			bsb.mutex.Unlock()
+			oldSkipList.DecrRef()
+			immutableBytes.Add(float64(-memSize), bsb.labelValues...)
 		}
 	}()
 	go func() {
 		defer bsb.writeWaitGroup.Done()
+		volume := 0
 		for op := range bsb.writeCh {
-			bsb.mutex.Lock()
-			if bsb.mutable.MemSize() >= int64(bsb.flushSize) {
-				select {
-				case bsb.flushCh <- flushEvent{data: bsb.mutable}:
-				default:
-				}
-				bsb.swap()
+			k := y.KeyWithTs(op.key, op.epoch)
+			v := y.ValueStruct{Value: op.value}
+			volume += len(k) + int(v.EncodedSize()) + skl.MaxNodeSize + nodeAlign
+			memSize := bsb.mutable.MemSize()
+			mutableBytes.Set(float64(memSize), bsb.labelValues...)
+			if volume >= bsb.capacity || memSize >= int64(bsb.capacity) {
+				bsb.triggerFlushing()
+				volume = 0
 			}
-			bsb.mutex.Unlock()
-			bsb.mutable.Put(y.KeyWithTs(op.key, op.epoch), y.ValueStruct{Value: op.value, Version: op.epoch})
+			bsb.mutable.Put(k, v)
 		}
 	}()
 }
 
+func (bsb *bufferShardBucket) triggerFlushing() {
+	for {
+		select {
+		case bsb.flushCh <- flushEvent{data: bsb.mutable}:
+			bsb.mutex.Lock()
+			defer bsb.mutex.Unlock()
+			bsb.swap()
+			return
+		default:
+		}
+		time.Sleep(10 * time.Second)
+	}
+}
+
 func (bsb *bufferShardBucket) swap() {
 	bsb.immutables = append(bsb.immutables, bsb.mutable)
-	bsb.mutable = skl.NewSkiplist(int64(bsb.size))
+	bsb.mutable = skl.NewSkiplist(int64(bsb.capacity))
+	immutableBytes.Add(float64(bsb.mutable.MemSize()), bsb.labelValues...)
 }